feat:상수,타입 분리(#336)
